Flyttet state for klage til modal for klage for å kunne legge til revurdering og håndtere state adskilt
